+++ sdev/include/shaiya/include/UserHelper.h
@@ -10,14 +10,14 @@ namespace shaiya
 
     struct UserHelper
     {
-        static bool ItemCreate(CUser* user, ItemInfo* itemInfo, uint8_t count, int& outBag, int& outSlot);
-        static bool ItemRemove(CUser* user, uint8_t bag, uint8_t slot, uint8_t count);
-        static bool ItemRemove(CUser* user, ItemInfo* itemInfo, uint8_t count);
-        static bool ItemRemove(CUser* user, ItemEffect itemEffect, uint8_t count);
+        static bool ItemCreate(CUser* user, ItemInfo* itemInfo, uint count, int& outBag, int& outSlot);
+        static bool ItemRemove(CUser* user, uint bag, uint slot, uint count);
+        static bool ItemRemove(CUser* user, ItemInfo* itemInfo, uint count);
+        static bool ItemRemove(CUser* user, ItemEffect itemEffect, uint count);
 
-        static void SetMovePosition(CUser* user, uint16_t mapId, float x, float y, float z, int movePosType, uint delay);
-        static void SetMovePosition(CUser* user, uint16_t mapId, SVector* pos, int movePosType, uint delay);
-        static bool Move(CUser* user, uint16_t mapId, float x, float y, float z, int movePosType, uint delay);
-        static bool Move(CUser* user, uint16_t mapId, SVector* pos, int movePosType, uint delay);
+        static void SetMovePosition(CUser* user, uint mapId, float x, float y, float z, int movePosType, uint delay);
+        static void SetMovePosition(CUser* user, uint mapId, SVector* pos, int movePosType, uint delay);
+        static bool Move(CUser* user, uint mapId, float x, float y, float z, int movePosType, uint delay);
+        static bool Move(CUser* user, uint mapId, SVector* pos, int movePosType, uint delay);
     };
 }
